requirements.txt
Django==4.2.19
import-linter==2.1
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.2
black==24.8.0
gunicorn==21.2.0
django-crequest==2018.5.11
django-auto-prefetch==1.9.0
psycopg2==2.9.9
pillow==10.4.0
python-dotenv==1.0.0
pytest-django==4.10.0
drf-yasg==1.21.10 
celery==5.5.0
django-crequest==2018.5.11
django-cors-headers==4.7.0
Faker==37.1.0
pyotp==2.9.0
django-redis==5.4.0